requirements.txt
requests
beautifulsoup4
pillow
django
sqlparse